+++ pettingzoo-unity/pettingzoo_unity/envs/unity_base_env.py
@@ -1,10 +1,8 @@
 import atexit
 from typing import Optional, List, Set, Dict, Any, Tuple
-
 import numpy as np
 from gym import error, spaces
 from mlagents_envs.base_env import BaseEnv, ActionTuple
-
 from pettingzoo_unity.envs.env_helpers import _agent_id_to_behavior, _unwrap_batch_steps
